ingestion/requirements.txt
yfinance>=0.2.50
requests==2.31.0
google-cloud-bigquery==3.17.2
